deps: bump tree-sitter to >=0.26.3 (resolves ast-grep-core conflict)
Workspace-root `Cargo.toml` pin raised from `"0.25"` → `"0.26"`, which
resolves to tree-sitter 0.26.8. The conflict: ast-grep-core >=0.40 requires
`tree-sitter ^0.26.3` (the `links = "tree-sitter"` key enforces a
single resolved copy), so both must land in the same series.

Root cause of the original pin was a misread of the grammar crate deps:
tree-sitter-rust 0.24.2, tree-sitter-python 0.25.0, and
tree-sitter-typescript 0.23.2 all declare `tree-sitter ^0.24/^0.25` only
as a *dev-dependency* (used solely in their own test suites). Their
production dep is `tree-sitter-language 0.1`, a semver-stable ABI bridge
crate with no `links` key — so upgrading the workspace runtime is safe.

One API break fixed: `Node::child(i)` now takes `u32` instead of `usize`
in tree-sitter 0.26. `read_children` in sittir-core/src/read_node.rs
cast `child_count` to `u32` and dropped the redundant `i as u32` on
`field_name_for_child`. The stale test in tests/read_node.rs was calling
`read_node` with 4 args (pre-existing bug, counter ref removed in an
earlier refactor) and asserting contiguous 0-based node IDs (counter
semantics, not pointer-id semantics); updated to the current 3-arg API
and relaxed to a uniqueness assertion.

Before/After (TS == native): pass 3172/4968, fail 1796 (unchanged).
regression-checker: PASS (baselines not touched, no JS test regressions).

+++ rust/crates/sittir-core/src/read_node.rs
@@ -147,13 +147,13 @@ fn read_children(
 ) -> (Option<HashMap<String, FieldValue>>, Option<Vec<NodeData>>) {
     let mut fields_acc: HashMap<String, Vec<NodeData>> = HashMap::new();
     let mut children_acc: Vec<NodeData> = Vec::new();
-    let child_count = node.child_count();
+    let child_count = node.child_count() as u32;
     for i in 0..child_count {
         let child = match node.child(i) {
             Some(c) => c,
             None => continue,
         };
-        let field_name = node.field_name_for_child(i as u32);
+        let field_name = node.field_name_for_child(i);
         let data = read_ts_node(child, source);
         match field_name {
             Some(name) => fields_acc.entry(name.to_string()).or_default().push(data),
